project/3c/scrapper-service/requirements.txt
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.0
selenium
epicstore_api
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
flasgger==0.9.7.1
